writingprompt_BM/requirements_eval.txt
rouge-score==0.1.2
bert-score==0.3.13
nltk==3.8.1
numpy==1.24.3
mauve-text==0.3.0
torch>=1.9.0
transformers>=4.20.0 